Small bug fix, minor code cleanup
>Replace $this with self in Utility class usage
>Further utilizatoin of Tag\Helper class
>call Utility::tag_load_config() (moved from /include/functions.php
>deprecate ./include/onupdate.php/tableExists() function - not used
>deprecate ./include/action.module.php functions - module uses
onupdate/oninstall/onunistall file(s) instead.
>changed module copyright to use xoopsmicrobutton.gif instead of
"missing" graphic
>deprecate ./include/formtag.php (XoopsFormTag class)
>deprecate ./include/functions.ini.php - use Utility class methods
instead
>deprecate .include/functions.php - use designated class methods instead
>add TagHandler::getTagData() method
>minor language file updates
>use header template in help rendering

+++ plugin/randomquote.php
@@ -8,18 +8,18 @@
  but WITHOUT ANY WARRANTY; without even the implied warranty of
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 */
-
 /**
  * XOOPS tag management module
  *
- * @package        tag
+ * @package        \XoopsModules\Tag
  * @copyright      {@link http://sourceforge.net/projects/xoops/ The XOOPS Project}
  * @license        {@link http://www.fsf.org/copyleft/gpl.html GNU public license}
  * @author         Taiwen Jiang <phppp@users.sourceforge.net>
- * @author         ZySpec <owners@zyspec.com>
+ * @author         ZySpec <zyspec@yahoo.com>
  * @since          2.33
  */
 
+use Xmf\Request;
 use XoopsModules\Randomquote\Constants;
 
 defined('XOOPS_ROOT_PATH') || die('Restricted access');
@@ -39,7 +39,14 @@ function randomquote_tag_iteminfo(&$items)
 {
     $items_id = [];
     $cats_id  = [];
-
+    /*
+    $cats_id = array_map('intval', array_keys($items));
+    foreach ($cats_id as $cat_id => $itemArray) {
+        $items_id = array_merge($items_id, $itemArray);
+    }
+    $items_id = array_map('intval', $items_id);
+    $items_id = array_unique($items_id); // remove duplicate item_ids if they exist
+    */
     foreach (array_keys($items) as $cat_id) {
         $cats_id[] = (int)$cat_id;
         foreach (array_keys($items[$cat_id]) as $item_id) {
@@ -58,18 +65,17 @@ function randomquote_tag_iteminfo(&$items)
 
     foreach ($cats_id as $cat_id) {
         foreach ($items_id as $item_id) {
-            $quoteObj                 = $quoteObjs[$item_id];
+            $quoteObj = $quoteObjs[$item_id];
             $items[$cat_id][$item_id] = [
-                'title'   => $quoteObj,
-                //                                                "uid" => $quoteObj->getVar("uid"),
-                'link'    => "index.php?id={$item_id}",
-                'time'    => strtotime($quoteObj->getVar('create_date')),
-                //                                               "tags" => tag_parse_tag($quoteObj->getVar("item_tag", "n")), // optional
+                'title' => $quoteObj,
+                //'uid' => $quoteObj->getVar("uid"),
+                'link' => "index.php?id={$item_id}",
+                'time' => strtotime($quoteObj->getVar('create_date')),
+                //'tags' => \XoopsModules\Tag\Utility::tag_parse_tag($quoteObj->getVar("item_tag", "n")), // optional
                 'content' => '',
             ];
         }
     }
-
     unset($items_obj);
 
     return true;
@@ -88,30 +94,22 @@ function randomquote_tag_synchronization($mid)
     /** @var \XoopsModules\Tag\LinkHandler $itemHandler */
     $linkHandler = \XoopsModules\Tag\Helper::getInstance()->getHandler('Link');
 
-    if (!$itemHandler || !$linkHandler) {
-        $result = false;
-    } else {
-        //        $mid           = XoopsFilterInput::clean($mid, 'INT');
-        $mid           = \Xmf\Request::getInt('mid');
+    $result = false;
+    if ($itemHandler && $linkHandler) {
+        $mid = Request::getInt('mid', 0);
         $moduleHandler = xoops_getHandler('module');
-        $rqModule      = XoopsModule::getByDirname('randomquote');
+        $rqModule = XoopsModule::getByDirname('randomquote');
 
         // check to make sure module is active and trying to sync randomquote
         if (($rqModule instanceof XoopsModule) && $rqModule->isactive() && ($rqModule->mid() == $mid)) {
             // clear tag-item links
-            $sql    = "DELETE FROM {$linkHandler->table}"
-                      . " WHERE tag_modid = {$mid}"
-                      . '    AND '
-                      . '    (tag_itemid NOT IN '
-                      . "        (SELECT DISTINCT {$itemHandler->keyName} "
-                      . "           FROM {$itemHandler->table} "
-                      . "           WHERE {$itemHandler->table}.quote_status = "
-                      . Constants::STATUS_ONLINE
-                      . '        )'
-                      . '    )';
+            $sql = "DELETE FROM {$linkHandler->table}"
+                 . " WHERE tag_modid = {$mid}"
+                 . " AND (tag_itemid NOT IN"
+                 . " (SELECT DISTINCT {$itemHandler->keyName}"
+                 . " FROM {$itemHandler->table} WHERE {$itemHandler->table}.quote_status = "
+                 . Constants::STATUS_ONLINE . "))";
             $result = $linkHandler->db->queryF($sql);
-        } else {
-            $result = false;
         }
     }
